feat: add freeze of attitude when multirotor is on the ground
